feature/train_test_split: split train/test on csv and adopt the change on tfrecord
