+++ py_app/criteo_ads_data/run_processing.py
@@ -14,31 +14,29 @@ if __name__ == '__main__':
 
     # arg parser
     parser = argparse.ArgumentParser()
-    parser.add_argument('--input_path', type=str,
-                        default='../../data/criteo_ads_data/dac/train_csv_1000000')
-    parser.add_argument('--output_path', type=str,
-                        default='../../data/criteo_ads_data/dac/train_tfrecord_1000000')
+    parser.add_argument('--input_path', type=str)
+    parser.add_argument('--output_path', type=str)
     args = parser.parse_args()
 
     # init config
     input_path = args.input_path
     output_path = args.output_path
-    input_files = glob.glob(f'{input_path}/*.txt')
+    input_files = glob.glob(f'{input_path}/*/*.txt')
     output_files = [
         x.replace('.txt', '.tfrecord.gz').replace(input_path, output_path)
         for x in input_files
     ]
 
-    # create output dir
-    dest_pathname = os.path.join(output_path, '')
-    if not os.path.exists(os.path.dirname(dest_pathname)):
-        os.makedirs(os.path.dirname(dest_pathname))
-        logging.info(f'Created directory {dest_pathname}.')
-
     # write .tfrecord
     logging.info(f'Loaded: {input_files}')
     for i_file, o_file in zip(input_files, output_files):
 
+        # create output dir
+        if not os.path.exists(os.path.dirname(o_file)):
+            os.makedirs(os.path.dirname(o_file))
+            logging.info(f'Created directory {o_file}.')
+
+        # write .tfrecord
         start_time = time.time()
         logging.info('Writing .tfrecord')
         tfrecord_writer(
